Converted TinyMCE plugin to a class
